First commit
data structures put in place. Seqnames and strand as rle encoded, ranges
as vec[start, end, width] for each of the rle encoded seqnames and
strand. Tested. Also added string to enum trait for seqnames and strand,
so just need to iterate over a polars dt and parse the seqnames and
strand.